create main forum automatically

+++ src/model/forum.rs
@@ -1,8 +1,8 @@
 #[derive(Clone)]
 pub struct Forum {
-    pub id: i64,
+    pub id: i32,
     pub title: String,
     pub description: String,
     pub slug: String,
-    pub category_id: i64,
+    pub category_id: i32,
 }
